refactor: move sha2 hint
